Added single Redis instance Redlock support.
The definition of the Redlock (template) class is available in
src/sw/redis++/redlock.*

The tests are implemented in test/src/sw/redis++/redlock_test.h*

The implementation is meant for a single Redis instance or for a
single clustered Redis instance. The locking of multiple instances is
not supported, but could be added.