Made RedLockMutexVessel usable with Redis cluster.
Up to now, a Redis reference had to be passed to RedLockMutexVessel at
construction time. This forced the user to pass the exact redis
instance that contains the key name, when using a cluster. If this was
not done, the cluster might tell the client that the key moved to
another node. As the Redis class does not know how to handle this, it
will result in an exception.

This commit turns RedLockMutexVessel into a template, that enables
the constructor to accept either a Redis or RedisCluster. This makes
the use of RedLockMutexVessel with or without clustering completely
transparent.