Keep connection on timeout
When reconnect is done due to timeout the output buffer is
lost, and it is unknown whether commands sent will or will
not reach the server, which can result in lost data. By
instead keeping the connection and receiving and ignoring
stale messages it is ensured at least that all commands
are sent in time.