Made it possible to run test/test_redis++ in parallel.
Running more than one test/test_redis++ in parallel on a single Redis
database/node, did not work, as the test key names were always the
same. This patch enables test/test_redis++ to be called with a '-d
db_name' option, for overriding the key's db-name. Furthermore, the
'-f' option was added for enabling the "SCRIPT FLUSH" test.

Note that the "SCRIPT FLUSH" test will now not be run by default. One
has to supply the '-f' option to force this test to be run.

This patch paves the way for some planned Docker images, that would be
able to test builds with different C++ standards, in parallel, using
only a single Redis server.