Changes for Nvidia compiler chain
Compiling redis++ with the Nvidia compiler chain fails due to warnings
thrown during compilation which are converted to errors because of the
`-Wall` compilation flag. Nvidia's compilers are based on the PGI
compilers which can be quite pedantic. In this case, a number of blocks
of code are seen as unreachable either due to a 'break' immediately
following a 'throw' or logic (in the case of reply.h).

The changes in this commit lead to a clean compilation of Redis++.
Some of these changes lead to some code that is more verbose than
is strictly needed, however are necessary to avoid warnings.